Make scale factor a double.

+++ src/main/java/com/jug/util/imglib2/Imglib2Utils.java
@@ -132,9 +132,9 @@ public class Imglib2Utils {
 //        ImagePlus tmp_image = ImageJFunctions.wrap(imgResult, "imgResults");
 //        IJ.saveAsTiff(tmp_image, outputFile.getAbsolutePath());
 
-    public <T extends NumericType<T> & NativeType<T>> Img<T> scaleImage(Img<T> img, int scaleFactor){
+    public <T extends NumericType<T> & NativeType<T>> Img<T> scaleImage(Img<T> img, double scaleFactor){
         Scale2D scalingTransform = new Scale2D(scaleFactor, scaleFactor);
-        FinalInterval biggerInterval = new FinalInterval( Arrays.stream( Intervals.dimensionsAsLongArray(img)).map(x -> x * scaleFactor ).toArray());
+        FinalInterval biggerInterval = new FinalInterval( Arrays.stream( Intervals.dimensionsAsLongArray(img)).map(x -> (long)(x * scaleFactor) ).toArray());
 
     //        BSplineCoefficientsInterpolatorFactory<T,DoubleType> interp = new BSplineCoefficientsInterpolatorFactory<>(img);
     //        InterpolatorFactory interp = new NearestNeighborInterpolatorFactory();
